Narrow down ProblemReporter API
Leave only one overload without 'line', leave a comment that its
usage is discouraged. For most cases when reportNodeError should be
preferred:

- even now, reportNodeError includes 'line' in the BuildProblem,
enabling better highlighting in IDE

- if all reports go through Node-based API, in future it will be
easier to implement even better diagnostics positioning strategies
(e.g. report diagnostic not on the whole line, but only on the
specific part of it, etc.)