[parser] Replace VirtualFile.toPath hack with toNioPath call
The hack was initially introduced as a workaround for using the light test fixtures in the plugin. The problem with the light test fixtures is that they use a temporary file system (TempFileSystem), which doesn't support conversion to java.nio.Path and just throws an exception.
However, using the hack which got the real java.io.File is also incorrect as it might return some nonsense for such temporary file systems. The better way for now is to use the heavier test fixtures (CodeInsightFixtureTestCase), which use the real FS underneath.