[schema] Rework traces and build problems API
- Traces change
Traces being of type Any isn't quite obvious to tooling. As we control the way AOM is created, we are aware of all the possible sources for model properties. Thus, a sealed hierarchy has been introduced.

- Build problem sources
Different clients might want to render the problems differently. While CLI wants to have line indicators and offsets, IntelliJ is more interested in having the corresponding PSI element. As we can report errors on PSI elements, it makes sense to avoid conversion to offsets and back to PSI by providing the means to receive the PSI element directly from the build problem.
Another improvement here comes with the BuildProblem now containing the range, not the line + column, allowing us to be more fine-grained with the reporting on the CLI.