[test] Fix include build path
Currently the settings.gradle.kts file of these e2e test projects
is completely ignored during the tests, because these test projects
are copied and the settings.gradle.kts is entirely replaced.

The includeBuild(...) call is only useful when interacting with
the project by hand (running commands from that dir). Since the move
of gradle settings to the root of our repo, we need to go up 5
levels (and not longer 4).