[test] Fix integration tests
The settings.gradle.kts was entirely replaced in test projects,
which made it impossible to test things related to the settings
themselves.

Now, we only remove specific lines from this file if they are
made redundant by adding Amper on the classpath of the build
(marked with <REMOVE_LINE_IF_RUN_WITH_PLUGIN_CLASSPATH>).
This is typically the case for the includeBuild(../../../../..)
which is only here so that manual tests in the test project
directory can work.