[dr] Use Job instead of CoroutineScope, and cleanup for gc
There is no need to use a whole CoroutineScope, as we just want a
handle for cancellation, which is what a coroutine Job represents.

Also, there is currently a race condition because we don't set the
scope (nor job) immediately upon creation in the nodeCache. Using
the result of launch makes it possible to guarantee this.

Last but not least, cleaning up the link to the job once the job is
done makes this instance eligible for GC and possibly reduces memory
pressure.