[dr] Refactor conflict resolution to prevent potential races
So far we used concurrent data structures to store information about
conflicts, but we accessed these structures multiple times in sequence,
making it possible for the state to be changed in the meantime.
Given the amount of things we check and change in these few lines, it
is safer to just wrap the whole block in a key-based mutex.

This also has the benefit of lifting the thread-safety constraint on the
data structures we use, and thus we can operate with a lock-free
coroutine mutex instead, which might give us a bit more performance as
well.