[dr] Reuse MavenDependencyNodes to avoid exponential work
MavenDependencyNodes represent a "desired" version and are not strictly
linked to a particular dependency (due to potential conflict resolution).
If a conflict resolution occurs, the dependency corresponding to a node
may have a different version from the "desired" version of the node.

We can theoretically use a tree (strictly speaking) of nodes without
reusing a node, even when 2 dependencies depend on the same 3rd one. We
can map these nodes to a _graph_ of dependencies, and still avoid resolving
the same dependency twice. That being said, for huge graphs with lots of
subgraph reuse, creating this giant tree might be too much. We can improve
this by reusing the nodes themselves too, which is what this commit does.

Because of this, nodes can now have several parents, and thus the graph is
no longer a strict tree.