[frontend] Replace jvm.target/java.source settings with jvm.release
The target and source settings are incomplete, and the use case of
cross-compilation to other JDKs is better served with the release
compiler argument, both in Kotlin and Java compilers.

Having a default means that, in Gradle-based Amper, we ignore the
user's choice of JDK, which is strange. But Android projects
require us to set source/target/release-related settings otherwise
there is a mismatch between java tasks and kotlin tasks which fails
the build. This is why we still keep the default release 17 for now.

We can revisit this later.