[gradle] Support Gradle > 8.6 for bundled Compose version + graceful degradation
The current Gradle version limitation (<= 8.6) is due to the API used to
apply the Compose plugin dynamically.
However, we do have a bundled version of the Compose plugin that matches
the default Compose version, so we can apply this Compose plugin without
the weird dynamic plugins shenanigans when the user doesn't customize the
version.