[frontend] Reduce API surface in ModelInit/SchemaBasedModelImport
We currently have the same API 4 times:
* in ModelInit
* in ModelInit.Companion
* in SchemaBasedModelImport
* in SchemaBasedModelImport.Companion

Only very few of those are used:
* The IDE only directly uses the companion object of SchemaBasedModelImport, and it's the only one which needs getModule/getTemplate
* The ModelInit abstraction with service loader is only used for its
getModel() function, and there is no need for other functions.

The repetition between instance methods and the companion objects seems to be due
to the impossibility to use Kotlin objects via a ServiceLoader. We can use a proxy
class to solve this without duplicating everything in the companion.

The non-blocking variant of the SchemaBasedModelImport's companion functions
are actually not used, and will probably never be necessary. In the IDE,
getModel is already used in a blockingContext, and getModule/getTemplate
cannot switch to a suspend approach right now because of where they are used.
They will need to be reworked anyway, as they are used only for their
diagnostics. In the meantime, let's remove the unused functions.