[backend] AMPER-393 Pass proper project root to delegated Android Gradle build
So far, we passed the module directory as a project root. This means
that, during model import, we considered the root of that module as
the root of the project, which may have unintended consequences.

For example, version catalogs might not be detected properly (too high
in the file system hierarchy). Internal module dependencies might also
not work.