[test] Pre-install the latest android cmdline-tools to avoid re-download
The android tests always want to download the latest cmdline-tools, despite
the fact that we already download a fixed version.
This creates a `$ANDROID_HOME/cmdline-tools/latest` directory which breaks
our incrementality. For this reason, we delete this directory before the
incremental cache check, but that means the tests have to download it again.

To avoid this, we now pre-install the latest version of the cmdline-tools,
and the tests can just use it. We also don't need anymore to delete this
directory for the sake of the incrementality cache, which is a double win.