[test] Pre-create lock files instead of deleting them to avoid AccessDeniedException
The lock files in the Android home can be created by the Amper CLI that we run in the
tests. Because of this, the "clean" Android home installation that we perform before
tests doesn't have the same contents as when we start tests a second time, and this
leads to 100% cache miss.
The previous solution to this problem was to delete leftover lock files when running
the tests again, just before the incremental cache check, to bring the directory back
to its initial cached state. The problem is that those things are actually lock
files, and by definition can be locked by another process.
As a new solution, we pre-create the lock files if they don't exist, so they become
part of the cache. This way, we don't have to delete them and we don't risk access
issues.