[frontend] AMPER-814 Make publication opt-in by adding publishing.enabled
So far, all modules were implicitly publishable. This makes it very
annoying to publish just what we need, because we need to list the
(root) modules we want to publish explicitly in the CI.

Instead, with settings.publishing.enabled, we can decide for each
module whether they should be published. We also have an inspection
that ensures any published module only depends on published modules.