[frontend] AMPER-919 Do not fail when searching for catalogs outside a guessed project root
We need to stop using `Project.guessProjectDir` because it's unreliable
and incorrect in several situations. We relied on this information to fail
fast when guessing the versions catalog location, but it led to false
negatives because of the incorrect project root.

We now still go up the file tree to find catalogs (still a heuristic), but
we no longer fail if the starting point is not below the root, which is
unnecessary and breaks several valid cases because the root is wrong.