[test] Fix Android license test
The missing license in the "empty" Android SDK home depends on the
version of the cmdline-tools that we install.
A new version was released, and now cmdline-tools require a
different license ID to be accepted, so the test should check for
a different failure.

We need to ensure we really have a fresh empty Android SDK home
directory before running the tests, so we avoid putting them in the
persistent cache on the CI.