Don't let JGit access global gitconfig
We call JGit to calculate the snapshot of the current local state of the code. While doing this, JGit reads all relevant git config files and fails if it recognizes an unsupported parameter (e.g. `diff.algorithm=patience`). Such a parameter can come from a global or system gitconfig, and the easiest way to workaround the issue is to avoid reading global/system gitconfigs (since it is unnecessary for the task anyway).