[startup] setup telemetry from the very start of the app
Previously, the telemetry was set up only when we knew the logs directory where this telemetry should be written to. But it happens quite late during startup, in particular, after the discoverability of the project root folder. The activities which happen during this time take quite some time, so we want to measure them as well.

Since we don't know where to write the telemetry, we first set it up without specifying the logs directory; we specify it later when we know it. Since the BatchSpansProcessor exports spans once in 5 seconds, there is enough time to set up the log roots dir (otherwise something is very broken in the app, and we don't care about the telemetry).

Some commands (which don't have the backend, e.g. `init` or `tool`) don't have the log roots at all. In this case, the field is not being set for them, and telemetry is never written. However, at this point we are not interested in telemetry of these commands, so it is ok for the sake of simplicity.