Switch to the gradle enterprise server for OSS
We have an agreement with Gradle for ge.jetbrains.com, but we
don't for ge.labs.jb.gg. It's unclear whether we're allowed
to use the private instance for other projects than Space.

Like Amper, Kotlin is OSS and has a private CI on TC buildserver
and they use the public Gradle Enterprise from the agreement.