[frontend] Align settings.kotlin.serialization.format default with docs
The default didn't matter so far, because both the short and the long form
required the format to be specified (unless people used the weird
"serialization: {}" syntax, which probably never happened in practice).

Now that we have the version setting. It makes a lot of sense for people
to just customize the version without adding a format, in which case they
should really get 'none' as default format.