[test] Fix flaky compose-multiplatform-room test by using different schema locations for each platform
We get an exception from time to time in our tests:
JsonDecodingException: Expected start of the object '{', but had 'EOF' instead at path: $

This might be because multiple KSP tasks are running at the same time, and
Room's configuration contains the same schemaLocation for all platforms.
It might be that concurrently reading/deleting those files is breaking things.

This commit uses a different path for each KSP platform to avoid this issue.