[cli] Move the lines running Amper to the bottom of amper.bat
When running `amper.bat update`, the script reaches the line where
it calls Amper via java, then Amper actually runs, and then control
is returned to the script to get to the `exit /B ...`.

During the execution of Amper's update command, the script itself
is replaced. Surprisingly, this doesn't fail on Windows. However,
the script seems to return to the next line after java was called,
but it might be based on line numbers, not on the actual java line.

This means the new bat script will resume execution at some arbitrary
point that corresponds to the former `exit /B` but it could be anything.
This causes problems like running random code lower in the script.

This commit tries to limit this issue by placing the java command in
last position.