[cli] Use the SHA instead of URL in the flag file for dist/JRE downloads
The URL might change depending on env variables, but the contents is what
matters. If the correct SHA is present, then we can skip the download.

This also helps our tests, because the port in the local HTTP server
changes every time, so the download would be done every time too.