[test] Separate test caches for different purposes
At the moment the whole sharedTestCaches dir is used as an Amper
user cache root, which means other caches like Android SDK or Gradle
home are mixed with directories from the Amper cache, which could
mess things up.

This commit brings a dedicated subdir to use as Amper cache to
avoid potential issues.