[cli] Adjust retries and timeouts for curl and wget downloads in wrapper
Curl doesn't retry by default, wget does 20. We align both
to retry 5 times instead (to reduce flakiness on CI).

The default read timeout of wget is 900s (15min). We reduce
this to 120 to avoid waiting forever when there is an error.
This is about idle time of the connection when nothing is
downloaded, so it doesn't affect slow downloads.