[test] Do not hide Gradle connection exceptions
The Gradle tooling API might throw many different exceptions.
In the test infra, we used to swallow any exception if the test
was meant to fail. We should instead focus on `BuildException`,
which is the exception representing failed builds.
The others should crash the test no matter what the expected
build result is.