[test] Prevent wrapper tests from using the main Amper's AMPER_JAVA_HOME
When tests are run by Amper, the test Amper process will use the same JRE
as the Amper process running the tests, because it inherits the
AMPER_JAVA_HOME env var set by our own wrapper.

We generally don't want this in wrapper tests, because the whole point is
to test things about the JRE download. This change fixes this to allow the
tests to pass even when run using standalone Amper.