[cli] AMPER-4025 Use custom listeners + JUnit Console launcher instead of custom launcher
There are many problems with the custom launcher for now.
In particular, kotlin-reflect (see KT-44795) and coroutines DebugProbes issues.
Both require that reflect/coroutines is in the same classloader as the stdlib,
which is not the case for our custom launcher, which uses a custom classloader.

In addition, the classpath of the custom launcher leaks into the user's tests.
This means that the users may have a conflicting kotlin-stdlib in their tests,
or any other library that our launcher will use in the future.