[test] Use deserialized traces from OTLP JSON file in tests
So far we had a special socket exporter to transfer Java-serialized
traces during our tests. This means that the tested data was
potentially different from the production data.
Also, this added a bunch of production code that was only there for
tests.

We now deserialize JSON traces like our users would, and thus we
test behaviors that are closer to production.