commonizer: fix the logic of collecting platforms for commonization: AMPER-4018
For each module, for each fragment in this module: take those sets of platforms which are targets in this fragment && contain only native platforms && contain more than 1 platform.

Add an integration test for the commonizer task. Utilize the KONAN_DATA_DIR environment variable to specify where the commonized libraries should be generated, instead of the standard place which is the Amper cache folder. The Amper cache folder is persistent, while for this test we want a clean empty folder to test what the commonizer generates.