project setup: for Gradle variant of amper.git, replace the IntelliJ fork of kotlix-coroutines with the original
IntelliJ Platform dependencies transitively bring com.intellij.platform:kotlinx-coroutines-[core-jvm|debug] dependencies to the project. They conflict with the original kotlinx-coroutines libraries, but dependency resolution doesn't handle this conflict because it is a different artifact.

In the Gradle version of the project, the IntelliJ fork appears earlier on the classpath, and that leads to a NoSuchMethodError in tests. This hack deliberately replaces IntelliJ libraries with their original Kotlin counterparts (as if dependency resolution would work here).

Amper version of the project does not have this issue. But it is just a lucky coincidence, so we'll need to fix it separately.

AMPER-MR-1572