Avoid calling IntelliJCoroutines
As mentioned in a previous commit, there are conflicting dependencies in the project: the vanilla kotlinx-coroutines-1.9.0 and IntelliJ fork of kotlinx-coroutines-1.8.0-intellij-11. Using both of them leads to a binary incompatibility error (some code gets compiled with 1.9.0 but is executed against 1.8.0 and faces a NoSuchMethodError).

We don't need IntelliJ modifications of the coroutines in amper.git, so we can just exclude the dependency. However, when we work with the PSI, some methods (e.g. PsiFileImpl.getChildren()) call ProgressManager.checkCanceled(), and the default implementation of that method calls `IntelliJCoroutines` - a special class added to the IntelliJ coroutines. Naturally, if we exclude the library, the class is not found, and getChildren() fails.

To work around it, we substitute the default CoreProgressManager with our implementation with just a basic cancellation check. Apart from a different ensureActive() call, our implementation does not run possible "check cancelled" hooks, but we don't expect to have any of these in the scenarios when ProgressManager is used from Amper (in particular, without any GUI).

This is needed to update the IntelliJ Platform: AMPER-4171