Move StandaloneXcodeComponentManager initialization closer to the usage
Currently, it's always registered when initializing a mock IntelliJ application on macOS. And it will fail in case there is no Xcode installed on the user's machine. However, that is a requirement only for some tasks and isn't true in general.

The change moves the registration closer to the usage (into ManageXCodeProjectTask) and makes it lazy allowing, e.g., building JVM apps without the necessity to install Xcode.