[cli] AMPER-4081 Fix staggered stacktraces in TeamCity format
Exception stacktraces are encoded into teamcity service messages by printing
their stacktrace, which uses tab characters before each "at" line.
The output of the test process was piped to the Mordant terminal, which
transforms tabs into spaces based on a tab width and virtual tab stops.
This means that the tabs in the stacktrace in the service message are
converted to different numbers of spaces depending on their position, which
leads to these misaligned stacktraces.