[cli][update] Schedule amper.bat copy in a shutdown hook to limit races
We have to schedule a separate process to update amper.bat in some cases.
So far we scheduled that process immediately during the update command,
towards the end of the command. Despite the delay of 1s before that
process does anything, it still could try to replace amper.bat before the
file was released by the amper update command.

This might happen if some cleanup at the end of the update command
execution takes unusually long. To be a bit safer in this respect, we now
launch the external process in a shutdown hook, so we're sure that most
cleanup has already been done (apart from the cleanup that is itself done
in shutdown hooks).