Nullify the previous IntelliJ application after disposing it
Previously, if the MockProjectInitializer was called twice, we were first disposing the previous Application, and then creating a new one. CoreApplicationEnvironment in its constructor first initializes the FileTypeRegistry, and then creates the Application.

Since adb30d1 in the IntelliJ repository (released in 2025.1), the FileTypeRegistry() constructor began sending events throw the ApplicationManager.getApplication().getMessageBus(). Since we already disposed the previous application, but did not remove it from the ApplicationManager, that our previous and already disposed Application instance was used by the FileTypeRegistry, which led to the AlreadyDiposedException from the message bus.

The fix is easy: remove the Application from the ApplicationManager (set it to null).