[infra] Restore the wrapper memory bump (to handle crazy outputs from R8 errors)
THe Gradle wrapper itself usually needs very little memory compared to the daemon,
because it just calls Gradle and forwards the output.
However, in our case, some tests building for Android print a LOT of R8-related
errors which are full stacktraces, and the wrapper cannot keep up, so we need this
extra memory.