[cli] AMPER-4229 Fix terminal output of teamcity format
So far, we had 2 competing requirements:

1. the teamcity format needs to avoid the Mordant terminal
processing that replaces \t with spaces, because it messes
up with the 'at' lines in stacktraces (AMPER-4081).
2. the test process output needs to go through Mordant to
play nice with our task progress widgets

It turns out we can go through Mordant to respect the widgets
but still not process \t and ANSI codes by using Terminal.rawPrint.
This is why I changed PrintToTerminalProcessOutputListener to use
rawPrint instead, and went back to using this for all 3rd party
process outputs.