[infra] Fix local build incrementality
So far, we rebuilt build.properties on every local change, which
made the whole build invalid because 'core' was rebuilt as a result.

It turns out we don't need this to prevent incrementality bugs in
local builds. We actually can build a hash of the distribution after
fact by hashing the jars in the current classpath (when running
a SNAPSHOT version - likely from local sources).

See the comment for details.