[cli] Remove commit hash from AmperBuild.codeIdentifier and improve version banner
The commit hash is now redundant there, and is actually harmful.
If we have local changes, make a build, and then commit the local changes,
the actual code hasn't changed and we shouldn't invalidate the cache.
So far, because the commit hash was present, it would invalidate the cache.

For incremental cache invalidation purposes, the codeIdentifier should
really just identify the code being run.
In non-snapshot builds, the maven version is unique so it's enough.
In snapshot builds, the distribution hash is necessary and sufficient.

Because we remove the commit from codeIdentifier, we put it back in the
banner, and also add the date while we're at it.