[concurrency] withDoubleLock: Enforce that the inter- and intra-process locks are the same
This forces the mutex lock (providing exclusivity within the current JVM) is
based on the same identity as the inter-process lock (the lock file) by using
the file path as mutex key.
Technically, since this is a striped mutex, it enforces that 2 calls to
withDoubleLock with the same lock file path are properly synchronized within
the current JVM, but it also may enforce intra-process synchronization for
different paths if their hashes are the same.
This is the price to pay for not cleaning up mutexes.