[concurrency] Simplify and fix StripedMutex
So far, giving a stripe count that was not a power of 2
actually had the effect of using a stripe count of the
next power of 2. This doesn't respect the doc.
Instead, we now enforce that the stripe count is a
power of 2 and always respect it.

This also simplifies the implementation.