[processes] AMPER-4284 Wait for processes to terminate on Ctrl+C
So far, processes started by Amper were automatically requested to
terminate when Amper itself is requested to terminate with Ctrl+C.
That said, because we're doing it in our Amper shutdown hooks, we
want it to be fast and we don't wait for the child processes to
actually terminate.

This has the unfortunate consequence of cutting off the output
of whatever process we were running, which is sometimes useful.
For instance, when doing ./amper run on a Ktor server, users
are sometimes interested in the logging done by Ktor on shutdown,
and not waiting for the process means not printing its output.

Worst case scenario, the child process hangs, and makes Amper's
shutdown hook hang as well. In that case, it may actually be
desirable to hang because we clearly show that some process is
still there. If we don't wait, we just leave a zombie process
silently.