[infra] Fix cli metadata publication
Old versions of Amper check the maven-metadata.xml of the old 'cli' artifact
name, so we need to keep publishing to this artifactId.
This is because even if we publish files with both 'cli' and 'amper-cli' names,
only the maven-metadata.xml of the artifactId specified in the publishing
settings will be updated.