AMPER-4408 Do not report overrides of the default values
Locking the default values makes sense as it makes the build deterministic with respect to Amper updates which can bump some versions and change the defaults. While locking values from the template might also make sense, it's a more exotic use-case that can be postponed until a better UX is found.