[cli] Fix JVM startup telemetry
The JVM startup was part of the "Root" span, but the Root span started
at the 'mainStartTime', which is after the JVM is initialized.
We needed to either make the root span start at the beginning of the
JVM startup, or take the JVM startup span out of the root span.

The former is what was done here, to see the whole command as one
single span in analysis tools like Jaeger.

(cherry picked from commit 9294fd96138b346a2ef01dff18866b881ad476a8)