[cli] AMPER-4436 Add working directory option to `run` command and use the current dir by default
It's important for ./amper run to respect the working directory because the app's arguments may
contain relative paths. If people want to use `./amper run --root some/project -- <args>` to run
their app from anywhere, we have to give them a way to keep the current directory as working dir.
This is actually a reasonable default, and people can still override it with the new option.