[dr] AMPER-878 Introduce a way to find parent KMP library coordinates
Some symbols can be represented both in the common source set of the original library and in one of its child platform-specific variants. Thus, when trying to suggest a library to add to the dependency for an unresolved symbol, we'll get two suggestions: one for the original library and one for its platform-specific variant.

This commit introduces an API to find the original KMP library allowing to deduplicate those options.