[cli] Improve (and fix!) telemetry reporting for commands with short-circuiting exceptions
Special command options like `--help` or `--version` (or custom ones) use an exception-based
short-circuiting mechanism. It is also possible to use it ourselves explicitly, by using
subclasses of `CliktError` like `throw PrintMessage()`.
See https://ajalt.github.io/clikt/exceptions

These exceptions are then translated by Clikt's `CommandLineParser.main()` into a call to
exitProcess() with a proper status code.

Now, when telemetry comes to the mix, we face 2 problems:

1. any exception thrown inside a span `use {...}` block will mark this span as "failed"
   (it appears with a red mark in some visualization tools like Jaeger).
   This means that even a harmless `./amper -v` or `./amper -h` will look like a failure.
2. any `exitProcess()` call inside a span `use {...}` block will bypass the finally
   block that marks the end of the span, and thus the span will never be reported.
   This is what causes a missing `Root` span (among others).

We already inlined part of Clikt's "main" helpers for the sake of more precise telemetry,
and this time we need to also inline the part that handles CliktError to fix these issues.

We want to exit early, properly report all spans in all cases, mark spans as failed only
when there is an actual failure, and exit with the correct exit code depending on what
the CliktError is.
This is why a new exception is introduced: ExitProcessButCloseTelemetrySpansException.
This allows to remember the exit code for real failures, and properly exit the process
eventually. For cases where it's not an actual failure (exit code 0), we just let the
code keep running, because anyway the cleanup is all we have left, so calling
 exitProcess(0) would be redundant.