[compilation] call the incremental Java builder from the JvmCompileTask if the corresponding property is enabled: AMPER-4834
To enable, call Amper with -Dorg.jetbrains.amper.jic=true

Store the actual JVM compile output in the taskOutput/output subdirectory. JIC generates incremental compilation data that should be persistent between task invocations, but it is not the compiler output artifacts. The compiler output artifacts are stored now in another subdirectory `taskOutput/jic-data`.

Call the Java incremental compiler in a separate process.
The jvm-inc-builder calls the Java compiler in-process, by using the JavaCompiler instance available in the current JDK. If we start JIC in-process as well, then it will use the JDK that Amper is powered by, and that has two problems:
 * Amper does not use a JDK anymore, it uses JRE. And the JRE doesn't contain the JavaCompiler.
 * Soon, the user will be able to specify the JDK to use for building the module, and that JDK should be used by JIC as well.