[cli] Use NoopProblemReporter instead of SLF4JProblemReporterContext in Android Gradle builds
If there were errors in the project model, the Amper CLI fails
early and the Gradle build wouldn't even run.
If there are warnings in the project model, they are already
printed in the CLI output when reading the model, so we shouldn't
print them again from the Gradle build.