[frontend] Remove VersionsCatalogProvider and its usages
Finding a catalog from every file doesn't make sense.
If it ever made sense with Gradle-based Amper, it no longer
makes sense.
We can directly find the version catalog based on the root
of the project.

For degraded modes where we inspect only a single module file
or template, we will find the catalog once and use it. It's
also fine to cache it. By the way, this way of finding a
catalog is generally not correct, and these helpers will soon
be removed anyway.