[frontend] AMPER-4515 Read templates safely and report them instead
Replace the diagnostic on the merged tree with diagnosing the templates when reading the module for the first time.
It's still not a perfect solution—it'd be better to replace it with the common approach of marking Paths in schema as "should exist" and providing a general diagnostic.

Cover the functionality with tests to avoid regressions.