[cli] Fix log level for Maven publish
The Maven publication system logs a TON of things at debug level,
which is more like a trace level (e.g. the body of every HTTP request).
It produces 900MB of debug logs for Amper's own publication.
We don't want that even in debug.log, so we need to limit the
level to INFO for this system.