[test] Workaround sequenceOf() bug with Kotlin 2.2.0 vs 2.1.20 classpath issues
Kotlin 2.2.0 adds an overload of sequenceOf() taking a single
item, but we have an issue with the test runtime classpath that
puts Kotlin 2.1.20 on the classpath instead 2.2.0, thus we get a
NoSuchMethodError.
This hack forces the use of the vararg overload that existed
before, avoiding the runtime error.
We'll need to fix the real classpath issue soon, though.