[frontend] Fix NoSuchElementException when building the model with no fragments
Technically, it should be impossible to have a module without fragments at this stage, but NoOpProblemReporter doesn't keep track of fatal errors and, thus, building a model with it doesn't fail early.

Nevertheless, the exact place that throws a NoSuchElementException can be aligned to just get the model from the context available previously.