[cli] AMPER-4505 Rename property aliases in schema to "misnomers" to better convey the role
These things are possible names that users might search for instead of the
correct property name. They are used for completion and "did you mean"
error messages.

These misnomers are not aliases in the sense that they are not another
way to refer a property. They cannot be used in the config files instead
of the correct name.