[compilation] jvm-inc-builder: write class files to the output directory
JPS generates jar instead of classes, but currently Amper expects classes. In particular, there are some use cases (e.g. Compose HotReload, Spring HotReload) that currently work with classes in the output. There could be adjusted in the future, but at the moment we decided to write the compiler output as class files.

To solve this, we generate class files together with writing the jar file, which now serves only for internal JPS purposes.
We also remove class files that are not the output of the compilation anymore, e.g. if the source file has been deleted between build invocations.

We also have to separately remove generated class files when a full rebuild is requested. In the case of full rebuild, ZipOutputBuilder.deleteEntry() is not called, so we need to inject to the code specific for rebuild.