[frontend] AMPER-424 Support customizing the Kotlin compiler version
This adds support for settings.kotlin.version, which is used
both for the compiler and for the stdlib dependency.

We also remove the defaults for -language-version and
-api-version compiler arguments, because it makes more sense
to just respect the compiler version by default now.
This avoids all problems related to computing the proper
language version from a given compiler version.