[compilation] use the proper fragment for generated sources
We already check at the beginning of the task, that there is a single leaf JVM fragment that we actually work on. We should refer to this single jvm fragment for generating the annotation processor sources rather than invoking the obscure firstOrNull that produces a generally undefined result. In fact, it takes the `common` fragment, but that is less correct than the jvm one.