[frontend] AMPER-4599 Support libs.versions.toml at the root
So far, we only supported a Gradle-style version catalog located
at gradle/libs.versions.toml. This gave the impression of still
being tied to Gradle, and was not desirable.

We have designed Amper's own version catalog but it's not yet
implemented. In the meantime, this enables placing
libs.versions.toml at the root of the project.

This commit also provides a way to locate the catalog file
and provides a VersionCatalog type hierarchy to solve the
"isPhysical" problem with composite catalogs. This is
necessary so the IDE can properly handle the new location.