[tests] improve the generation of the test failure message
Previously, we were passing List<String> as expected and actual values. Therefore, in the output and for the Show Difference feature, these values were converted to a single String via the standard joining mechanism: [A, B, C]. For long values it was unreadable because it made two very long Strings.

After this change, we provide a single String as a value, but we join it via newlines. The output generator doesn't need to transform these objects to Strings anymore, and displays them as is: every item is on a separate lines. This is especially useful for TeamCity service messages.