[tests] improve the assertion that TeamCity service messages are equal
Previously, a message from the expected list was compared against that message from the actual list, that had the same position in the list. It worked well if the message data was incorrect (e.g. the assertion message or a stack trace). But if the service message was missing from the list, the assertion would compare two arbitrary messages that have nothing to do with each other.

Instead, let's use the existing assertEqualsWithDiff that compares all messages altogether, and generate a unified diff that can be useful on the CI.