[cli] Fix length check in update command
When deciding whether an in-place update would be harmful, it is
incorrect to compare the length of the old and new wrappers.
What we really want to know if whether the offset we will resume at
is before the end of the new wrapper. We don't resume at the end of
the old file, but a bit before the end, at the beginning of the
'exit' command. We need to account for that.

(cherry picked from commit 4bf069d3c5c9a7c79564ae9da3c51f02232c59c5)