[templates] update the compose-multiplatform template: AMPER-4724
The previous version of the template is imported into IntelliJ incorrectly due to AMPER-4723 (which should be fixed separately anyway). But apart from that, we also do not need these dependencies in the shared module.
Also, it is not necessary to use the AppCompatActivity, the regular ComponentActivity works well nowadays.

The updated template is also more in line with the sample at `examples/compose-multiplatform`.