[core] Reuse the same client for multiple downloads
We used to use the `.config { }` function from Ktor, which creates
a new HttpClient every time. This is not resource-efficient if
we have many downloads.

Also, the config for the client used to download archives was
partially made in the general httpClient that we use in other parts
of the Amper CLI to download some XML file descriptors. These
should definitely have different timeouts, so better to separate
them.