[compilation] Introduce settings.java.compileIncrementally to enable or disable incremental compilation of Java sources
It is disabled by default at the moment.

JIC is used if either this user option is enabled, or if the 'hidden' JVM option "-Dorg.jetbrains.amper.jic" is enabled.
But if there are Java annotation processors, then JIC is not used regardless of the options, because annotation processors are not supported by the JIC yet.